Change stage SSL path
This is because certbot picks one of the domains on the list for the
'root' folder when a key is shared. After remaking the key to remove
some old domains, the folder picked was different from last time.

This will go away as we generally move away from certbot and stop
shoving everything on a single server that can't be rebooted.